Service interfaces, repository primary key changed from long to integer

+++ src/main/java/cz/cvut/fel/rsp/ReservationSystem/dao/CashRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CashRepository extends JpaRepository<Cash, Long> {
+public interface CashRepository extends JpaRepository<Cash, Integer> {
 }
