+++ src/main/java/cz/cvut/fel/rsp/ReservationSystem/dao/CustomTimeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomTimeRepository extends JpaRepository<CustomTime, Long> {
+public interface CustomTimeRepository extends JpaRepository<CustomTime, Integer> {
 }
