+++ src/main/java/cz/cvut/fel/rsp/ReservationSystem/dao/IntervalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IntervalRepository extends JpaRepository<Interval, Long> {
+public interface IntervalRepository extends JpaRepository<Interval, Integer> {
 }
